Refactor and make better use of the video viewmodel with databinding
