+++ app/src/main/res/layout/activity_video_details.xml
@@ -7,8 +7,8 @@
 
     <data>
         <variable
-            name="workoutVideo"
-            type="com.example.workoutapp.models.WorkoutVideo"
+            name="workoutVideoViewModel"
+            type="com.example.workoutapp.ui.videos.WorkoutVideoViewModel"
             />
     </data>
 
@@ -73,7 +73,7 @@
             android:textColor="#000"
             android:textSize="24sp"
             android:textStyle="bold"
-            android:text="@{workoutVideo.title}"
+            android:text="@{workoutVideoViewModel.workoutVideo.title}"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintRight_toRightOf="parent"
             app:layout_constraintTop_toBottomOf="@id/workout_video_detail_image"
@@ -87,7 +87,7 @@
             android:paddingTop="5dp"
             android:paddingLeft="15dp"
             android:textSize="18sp"
-            android:text="@{workoutVideo.instructor}"
+            android:text="@{workoutVideoViewModel.workoutVideo.instructor}"
             android:textStyle="bold"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintTop_toBottomOf="@id/workout_video_details_title"
@@ -102,7 +102,7 @@
             android:paddingLeft="15dp"
             android:textSize="14sp"
             android:textStyle="bold"
-            android:text="@{workoutVideo.level.value}"
+            android:text="@{workoutVideoViewModel.workoutVideo.level.value}"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintTop_toBottomOf="@id/workout_video_details_instructor"
             tools:text="BEGINNER"
@@ -126,7 +126,7 @@
             android:paddingLeft="15dp"
             android:textSize="14sp"
             android:textStyle="bold"
-            android:text="@{ @string/rating(workoutVideo.rating) }"
+            android:text="@{ @string/rating(workoutVideoViewModel.workoutVideo.rating) }"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintTop_toBottomOf="@id/workout_video_details_title_divider"
             tools:text="99%"
@@ -151,7 +151,7 @@
             android:paddingLeft="35dp"
             android:textSize="14sp"
             android:textStyle="bold"
-            android:text="@{ @string/difficulty(workoutVideo.difficulty) }"
+            android:text="@{ @string/difficulty(workoutVideoViewModel.workoutVideo.difficulty) }"
             app:layout_constraintLeft_toRightOf="@id/video_details_thumbs_up"
             app:layout_constraintTop_toBottomOf="@id/workout_video_details_title_divider"
             tools:text="8.5/10"
@@ -165,7 +165,7 @@
             android:paddingTop="5dp"
             android:paddingLeft="15dp"
             android:paddingBottom="5dp"
-            android:text="@{ @string/number_of_ratings(workoutVideo.numberOfRatings) }"
+            android:text="@{ @string/number_of_ratings(workoutVideoViewModel.workoutVideo.numberOfRatings) }"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintTop_toBottomOf="@id/video_details_rating"
             tools:text="33,000"
@@ -223,7 +223,7 @@
             android:paddingTop="5dp"
             android:paddingLeft="15dp"
             android:textSize="16sp"
-            android:text="@{workoutVideo.workoutEquipment.value}"
+            android:text="@{workoutVideoViewModel.workoutVideo.workoutEquipment.value}"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintTop_toBottomOf="@id/workout_video_details_equipment_divider"
             tools:text="Treadmill"
@@ -235,7 +235,7 @@
             android:layout_height="wrap_content"
             android:paddingTop="5dp"
             android:paddingLeft="5dp"
-            android:src="@{workoutVideo.workoutEquipment.icon}"
+            android:src="@{workoutVideoViewModel.workoutVideo.workoutEquipment.icon}"
             app:layout_constraintLeft_toRightOf="@id/video_details_equipment"
             app:layout_constraintTop_toBottomOf="@id/workout_video_details_equipment_divider"
             />
@@ -268,7 +268,7 @@
             android:layout_height="wrap_content"
             android:paddingLeft="15dp"
             android:paddingRight="15dp"
-            android:text="@{workoutVideo.description}"
+            android:text="@{workoutVideoViewModel.workoutVideo.description}"
             app:layout_constraintLeft_toLeftOf="parent"
             app:layout_constraintTop_toBottomOf="@id/workout_video_details_description_divider"
             tools:text="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."
